+++ CHANGELOG.md
@@ -2,3 +2,18 @@
 title: 更新日志
 toc: hidden
 ---
+
+### 0.0.22
+
+`2020-04-04`
+
+- Feat
+  - `radio-item`增加属性`plain`镂空属性
+  - `drop-select-item`增加统一表单滑动选择器demo实例
+
+- Fix
+  - `drop-select-item`修复type为single时赋初始值不显示问题
+  - `drop-select-item`修复type为work时三级职业回显问题
+  - `check-group`修复itemObject参数
+  - `check`修复check选中icon参数
+  - `picker`修复picker demo显示问题